Lexer: Begin implementing, add Token & TokenType
Gonna try minimize the annoyance of keywords on users by only enforcing
keywords in scopes where they apply e.g. args or rad blocks.

Also modify the EBNF to allow for a decoupling of the arg variable name
RSL devs write, from what RAD users will need to input. The use case is
best demonstrated by the fact that 'regex' is a keyword in the args
block. If the RSL dev wanted to have a user-facing arg named 'regex',
then they couldn't, because that's a keyword and not allowed. So,
instead of doing

 regex r string # A regex.

which is disallowed due to the regex keyword, they can do

 input_regex "regex" r string # A regex.

input_regex becomes the name of the variable for the dev to work with,
but the RAD user will see 'regex' as the name of the arg to define.

The alternative is to try and be a bit smarter within the rad block
about when we enforce 'regex' being a keyword. In theory possible, but
let's just take this alternative, which is pretty versatile.

Another unrelated thing: we're gonna try to give the RSL dev feedback on
the particular character in the offending line which is the problem, so
we also count char index within the context of a line.