Lexer: Emit INDENT tokens
Okay, I think this model will allow me to properly handle indentation in
the parsing stage. Basically, if we are at the start of a new line, and
we see two spaces, then we begin to lex out a space INDENT token. We
count the number of spaces in the indent. Similarly, if we're at the
start of the line, and we see a tab, we begin reading tabs until we hit
something else, and emit a tab INDENT token with the number of tabs. In
either case, if we read a mixed space/tab indent, we error.

I'm then imagining, in the parsing stage, the first INDENT token we
parse which comes after a COLON (and NEWLINE), will become the reference
'INDENT' expectation for the rest of the script. If we read other
consequential INDENTs (i.e. they're after a COLON NEWLINE), which do not
match some multiple of the reference indent, then that's an error.
Unlike Python, I'm imagining we just take a pretty hard stance on not
mixing indentation levels, it's not hard to get right (famous last
words).