Lexer: Add json field paths support
Taking a simplified approach which doesn't require quotes, but just
escaping the meaningful characters such as '.' or '[' that would
otherwise end the element.

I am arguably pushing too much parsing logic down into the lexer. For
example, I am choosing to treat a path element *and the following
brackets []* as part of the same token, if the latter is present. It
'feels' wrong and like the lexer is overstepping its bounds this way,
but I don't have tangible arguments against it if it's so easily doable
at this stage and might simplify the parsing later.

Also correct some lexing on identifiers -- I had not been allowing
underscores or digits in the name.

With this change, we can now fully lex the starwars1.rad example, and
correctly, as far as I can tell.