Parser: Add ast.go generator, begin parser
Very much inspired by Crafting Interpreters, we take the approach of
code generating a lot of the repetitive structures relating to
Expressions, Statements, etc. I include a Makefile which has a 'all'
command which does the generation, formatting, and final 'actual' go
build.

We add the parser struct which will do all the parsing. Right now, it
just handles newlines and primary assignments.

I realized I am not emitting 'bool literal' tokens in the case of
true/false from the lexer. That's because I had initially modeled those
as keywords, which I was gonna push into the parser. I think the
exceptions to this are true/false tho. It doesn't depend on the context
-- those should be interpreted as literals, I think. Todo for later,
tho.

Also bump Go to 1.22 cause I noticed I wasn't on the latest 🤷