Lexer: Change indentation emitting
I was working in implementing block handling in the parser and realized
that my approach was gonna be pretty tricky. After some more online
reading, I've found another approach, described here:

 https://marcelgoh.ca/2019/04/14/scanning-spaces.html

I've found this algorithm referenced/described elsewhere too, but this
blog post did a good job of explaining it succinctly (though didn't seem
to handle dedents when back on the 'global' 0-length indent level,
unless I missed it).

Basically, it results in the lexer emitting INDENT/DEDENT tokens a
little closer to how { and } bracket tokens would be emitted if those
were the path we took. This means it should be easier for the parser to
handle blocks.

We'll see!