Parser - args block: parse arg declarations
Okay, I think the INDENT/DEDENT changes from the previous commit are
good, blocks are not too bad to parse. There's a bit of
'consumeNewlines' going around, but I think that's okay for now.

We require args to be declared at latest after the file header, near the
front. I'm not certain this is the right choice, but for now, it will
reduce the number of 'strange' scripts and eases handling things a bit.
The more I write this, the less convinced I am that this is something we
should keep long term tho :P

Regarding keywords, we commit to the approach of detecting if an
identifier is a keyword in the parser. I think this will work well
because we will handle 'global' statements separately from arg and rad
block statements, so we can check for the correct subset of keywords in
the correct contexts.