Begin interpreter
Rather than implementing the parser in one go, I'll also begin
implementing the interpreter so that we can begin getting some runnable
scripts, and checking that everything works from end to end.

This instance of Rad probably doesn't run (haven't tried), I'm
committing before a meaningfully runnable checkpoint. But it includes
the stub definition of the interpreter (and args interpreter), Env, and
some other things.

Next step will be to have cobra handle args and pass them into the
interpreter's environment before it runs, so that the interpreter can
then 'activate' the args when it comes across the args block.