EBNF: Add query, table blocks
I feel relatively good about this (maybe with the exception of the table
block, but idea overall is good). We basically have a 'query' block
which does the query-extract operations, and then a 'table' block which
will do the display, and 'rad' block which is basically a union of the
two, performing both actions at once, allowing concision when the
'all-in-one' is appropriate.