EBNF: Add function calls
This is an important one -- a lot of utilities I want to provide via RSL
will be done via functions, to try and avoid complicating the language
itself. The most basic of these is perhaps printing. I intend to have a
print function which can be used as e.g.:

 print("Input: {input}", color="yellow")

The first arg is positional, but then there will be a bunch of default
args that can be overriden e.g. color.

It remains to be seen how I'll implement these. I'm very unlikely to
implement actual functions in RSL, so these function calls will probably
just be parsed themselves in the parser, and I will implement them in
Go, running the functions in the interpreter.

I think my philosophy on what functions should be, is that they should
either *do* something non-stateful (e.g. printing), or they should
return something which can later be used or just thrown away (e.g.
lower). They should not have stateful side-effects. So for example, I
would not make the sort statement in the rad block a function, because
that is *setting* behavior for the containing rad block.