EBNF: Add array assignment declaration syntax
e.g.

 AudCosts float[] = [0.1, 0.2]

This also fits neatly into the arg default syntax, allowing us to use
this for args. We rework types EBNF a little in the process.

Also realized we need switch cases to be on literals, not primitives. We
don't want to allow e.g. IDENTIFIERs. This is consistent with other
languages e.g. Java.