WIP: Cobra script sub command integration
This commit is not quite complete, but is almost there. I've been
experimenting a little in another sandbox repo, trying to demo to myself
that I can finesse Cobra into doing what I want re: rad. Rad is a
somewhat unique usecase in that it can only tell Cobra things about how
args should be passed, *after* rad has had a chance to interpret the
input script (if given) and extracted the args from the script. That
last part is what is missing from this commit.

This commit just contains the types and general logic we'll try
applying. Not super sure how robust it is. If it proves unwieldy in
practice, we *might* need to override a lot more things, or perhaps even
throw away cobra entirely and do our own arg parsing.

Next commit should actually look at the parsed statements and extract
the args from the given script, and pass that to Cobra.