Lexer: Handle shebangs
If a line starts with #! , we interpret that as a shebang and throw away
the rest of the line. If we detect a shebang on not the first line, we
error (I tested this locally with python and non-line-1 also did not
work).

Fix up a bit of rewinding detail on the indentation lexing. A bit messy,
but can hopefully clean up later if I'm so inclined.

Alter the lexer error to point at the first char index instead of the
point where we realized something was invalid.