Remove nulls from RSL
I started implementing VisitBinaryExpr and it became evident just how
insane the implementation of things have become due to the pointers
everywhere, all for the sake of allowing nulls in the language.

Reflecting more on that, the only thing nulls really would be useful for
is making that the value of optional args if they weren't set. But in
reality, I can just have a 'set' flag on those args, and disallow them
being used if they're not set. If you really need to check if an arg is
set or not, we can have a isSet(arg) function.

This commit gets rid of a ton of pointers and simplifies things pretty
well, I think VisitBinaryExpr will be a *bit* more doable after this...