Implement rad json trie extraction
Another milestone: basic rad block working! With this commit, the
following:

 Name = json[].commit.author.name
 Email = json[].commit.author.email
 Date = json[].commit.author.date

 url = "https://api.github.com/repos/torvalds/linux/commits?per_page=2"

 rad url:
     fields Date, Name, Email

correctly prints:

 Querying URL: https://api.github.com/repos/torvalds/linux/commits?per_page=2
 DATE                    NAME            EMAIL
 2024-09-02T02:59:59Z    Linus Torvalds  torvalds@linux-foundation.org
 2024-09-01T21:43:08Z    Linus Torvalds  torvalds@linux-foundation.org

The algo for extracting is not conceptually complicated -- we create a
trie out of our defined json fields, and then use it to traverse the
json blob received in our request, writing down the data for where we're
at when we hit a 'leaf' json field definition as the user has defined
it.

We use github.com/olekukonko/tablewriter for formatting tables to CLI.
It seems to be the go-to, even if the project itself is not that active.
But also, sometimes things are just complete and can be left alone, so
not necessarily bad!