Implement switch statements
A bit complex! And not entirely done, there are still three things
missing:

1. fuzzy matching on case keys
2. true case blocks, with yield statements
3. multiple return values / identifier assignment

But let's focus on what's done here in this commit.
The following syntax now works:

finalBase = switch base:
    case "github": "api.github"
    case "gitlab": "not supported D:"

url = switch:
    case: "https://{finalBase}.com/repos/{repo}/commits?per_page={limit}"
    case: "https://{finalBase}.com/repos/{owner}/{project}/commits?per_page={limit}"

finalBase will be resolved based on the value of 'base', similar to
traditional switch statements from other languages.

url is a little more interesting -- we implement an algo in
interpreter_switch.go which picks the case which matches best by
checking the string interpolation. See code for comment with details.

One meaningful change required for the implementation to work is that
optional args are no longer added to the environment with the 'basic'
value that cobra gives, for when a flag isn't specified. This may or may
not need to change in the future.

We need to temporarily disable string interpolation in the
LiteralInterpreter for the sake of the interpolation-based case picking.
We use a stateful bool, which is a little hacky, but gets the job done.
Everything is single-threaded anyway, so as long as we're stringent
about putting this bool back to the state it was before, should be ok.