Allow multiple assignment for switch blocks
This was somewhat artificially disallowed before. I waaanted to allow
multiple assignment possible for non-switch statements, but implementing
it isn't trivial and also doesn't seem like a super beneficial feature
at this point. It might just be a little inconsistent. I can see a world
where it'd be useful for functions to return multiple values, so we
might need to figure it out at least for functions in the future.