Fail sooner if shorthand flag is more than 1 char
Multiple chars are not allowed for shorthand flags, in part so that you
can have bool flags 'a' and 'b' and then the user specify them both with
a simple `-ab`. It's just the convention for *nix systems, and it's also
enforced by the arg library cobra uses (spf13/pflag), so it currently
will fail via a panic from the pflag lib, but we can provide better
feedback if we catch it ourselves, so do!