Fix single-assignment switches, allow explicit types on vars, etc
A few little good things in this change.

We improve the assignment parsing to allow setting explicit types on
non-arg declarations, including arrays and non-arrays, and switch
assignments. We then put this information on the assignment AST, to then
be validated in the environment when the variable gets set.

As part of this, we erase ArrayAssign and merge it into PrimaryAssign,
which is a nice simplification.

Switch statements previously did not allow single-var assignment. This
was definitely an oversight, we fix that up.