Implement for-each loops
Syntax examples:

for v in a:
    print(v)

for i, v in a:
    print(i, v)

i.e. Go-inspired (or Python enumerate) syntax.

This commit also implements var definition environment scoping for
blocks, for both if statements and for loops, the latter which I forgot
in the previous commit. I'm not actually convinced scoping for if
statements makes sense in a scripting language like this -- it could be
less verbose to just be able to do

if a:
    b = "hi"

print(b)

and have that work? idk

continue/break statements aren't completely implemented yet, that's
next.