Implement for loop break/continue
Implementing these in Go is a litte treacherous... I think I've got it
right, but the only way I could see to do this without overhauling e.g.
every Stmt returning some sort of 'control' flag, is to set some
stateful flags on the interpreter itself. It *looks* correct, and some
testing with scripts seems to do the right thing, but I wouldn't be
terribly surprised if a bug around this is uncovered at some point...

Fix a silly bug in the lexer where I accidentally would not emit tokens
for identifiers starting with j.