Make bool args always default to false
If a bool is declared like so (example)

args:
    name string
    capitalize c bool

Then it's sorta silly to require it to be set each time. The user has a
few options for defining it, including (but not limited to)

 mycmd alice true
 mycmd alice 1
 mycmd alice -c

The first two options are offered by rad, but not super conventional.
The main way to think about bool args is presence of flag, or
lackthereof. So if the user is not specifying the flag, we should assume
false. Technically the RSL user could always write

 capitalize c bool = false

but that adds verbosity. We can heuristically make the decision that
this should just always default to false.