Pass File Header to Cobra for --help display
Before, we were lexing and parsing the file header, but not passing it
to Cobra to actually be used to help the user.

Let's change that and actually populate the short/long descriptions
appropriately.

One improved behavior I want is that currently, this is an error:

"""
Yadda yadda really great script.
Use it!
"""

This is disallowed because you're not allowed to have a block (long)
description without also a one-liner, but the two are expected to have a
newline between them. Really, this should just get interpreted as one
long description, and no short description, which I think is fine.