Implement reading scripts from stdin, and outputting bash export cmds
This is a cool feature, I think, and will make rad still helpful for
cases where RSL itself isn't the right tool (or the problem can't be
solved fully with RSL).

Basically, we add a couple of global flags:

 --STDIN
 --BASH

Notice they're both all-caps, the goal is to differentiate these global
'rad' flags from potential user flags, which we'll encourage to be
lowercase.

STDIN makes rad expect the RSL script code to come in via stdin, rather
than via a file path it needs to load.

BASH makes rad, after interpreting whatever script it's running, print
a bunch of bash-compatible 'export' statements, for all the variables
defined in the interpreter's environment at the end. If invoked from a
bash script, this output can then be eval'd, defining those same
variables in the bash environment, ready for use via e.g.
'$example'-style syntax.

This is great because it allows bash script authors to leverage rad's
powerful, declarative arg syntax and handling. Bash arg parsing quickly
becomes unwieldy once you're doing anything with flags/optional args,
and is one of the primary motivations for rad's existence. If rad can at
least make that easier, then that's success!

Here's a demo of what it looks like in bash script:

rsl="\
args:
  name string # Name of the person
  age int = 30 # Age of the person"

eval "$(rad --STDIN --BASH $* <<< "$rsl")"

echo "Name: $name"
echo "Age: $age"

We define the RSL code in a 'rsl' multi-line var. Then, we pass that
through stdin using a 'here string' (as the bash lingo calls the <<<
operator), to rad. We prepare rad to ready from stdin by setting the
--STDIN flag, and ask it for exports with --BASH. We wrap all that in a
bash eval statement to evaluate the output, evaluating the exports.

A big change I had to make for this to work was to revamp the way we
integrate cobra. Before, we'd create a subcommand and re-run the root
cmd to let cobra handle it all over again.
This was hard to get working in combo with STDIN because subcommands are
reliant on a non-empty string 'command name'.
The simplification we make, which works quite nicely, is simply to
modify the root command to match the metadata extracted from the RSL
script, and rerun the root. This is actually also nicer for script path
invocations.

Misc note:

We trim the UseString generation because the 'stdin' "command" doesn't
have a name, thus the use string ends up slightly to the right because
it starts with whitespace.