Print help to stderr, and output exit 0 when help invoked from bash
This improves bash embedding pretty significantly. Rather than letting
the bash script continuing after an unsuccessful rad invocation (or if
help was invoked), we now print the help to stdout, and then 'exit 0' to
stdin, so that the wrapping bash 'eval' call evaluates exit 0 and quits.

Also reword some of the comments around the SetHelpFunc hack, as it's
out of date.