Fix bug: remember modifications to existing vars made inside a block
Previously, the following would print "alice":

name = "alice"
if true:
    name = "bob"
print(name)

but clearly, the expectation from any programmer will be that the
modification made in the if-block modifies 'name' and that this will be
persistent after exiting the block.

This is not currently the case, until this commit. Simple fix it to make
the Environment's setters delegate to the enclosing environment if it
exists.