Fix bug: Make json extraction algo use Literal, not Lexeme
If we have a json field definition like this:

field = json.parents[].names

then 'json' is one path token, 'parents[]' is another, and 'names' is a
third.

If we look at 'parents[]', 'parents[]' is the lexeme, and 'parents' is
the literal, with '[]' stored as a separate nillable token.

In the json algo, when looking up keys in the json, I've been
incorrectly using the lexeme instead of the literal. The key is
'parents', not 'parents[]'.

This fixes that.