Further improve bash embedding
rad wasn't handling some aspects of how it was getting invoked from bash
super well. For example, it'd print errors to Stdout instead of Stderr,
causing bash evals to further fail.

With this commit, we change the --STDIN bool flag into a string flag,
which requires being given a name for the script it'll read from stdin.
This is to help with the help string, but also to get around an issue
where cobra complains when the command doesn't have a name as is getting
executed.

Change the BASH flag to SHELL, I don't mean to only support bash.

We get the basename of the script path now, including when bash passes
"$0". It just looks better and more correct in the help string.

We augment the initial script arg reconciliation with a bool
determining if we should just print the help string an exit. The goal is
to print and exit if any of the required, non-default args are provided,
either via flag or positional.

We also will begin to hide some of the global rad flags that are
irrelevant to the particular script. Longer term, we might customize how
the help gets printed, and instead of hiding them entirely, print them
under its own section.

Lastly, if we error in cobra_root.go, write 'exit 1' if in shell mode,
so that it can get eval'd and the bash script can exit also.