Allow int shorthand flags, fix arg renames, improve help print
-- Allow int shorthand flags --

Previously, you could not have a shorthand flag which was an int. That
was because the lexer outputs it as a INT_LITERAL, but I am expecting
IDENTIFIERs (which works for non-int shorthand flags). Let's handle
both.

-- Fix arg renames --

I was erroneously using the Lexeme, not the Literal (again) for arg
renames, so the double quotes would get included in how users should
functionally set them. Not great. We correct that.

-- Improve help print --

We make bool flags always appear as optional, because they're either
explicitly defaulted to true/false or implicitly defaulted to false.
Also, we print them with prefixing -- to indicate more clearly they're a
flag in the usage string.

---

I also make all the ./tests/* into executable scripts, most redundantly,
but may as well.