Introduce 'printer' funnel for all rad output & new output flags
Ideally I'd done this from the start, but better late than never.

This commit introduces a 'Printer' interface/instance through which
*all* output from rad and its scripts travels. This allows the
introduction of some new global flags, which impact all these outputs.

--QUIET:     to suppress 'regular' output
--DEBUG:     to allow debug logs for RSL devs
--RAD-DEBUG: to enable debug logging for RAD internals

I'm not confident in how I've chosen to make rad behave when multiple of
these flags are enabled, as they overlap a bit, but I've doc'd it in
printer.go.

Also, I should revisit the impl a little as I'm currently passing the
printer instance around everywhere, which is a bit excessive in places
e.g. in json_algo.go, the Nodes and Trie have to be given the instance
so they can log errors in the algo...

Still a few straggling places where we panic, not through the printer.
They're not that pressing, or likely to be hit, though.