Improve table width-shrinking algorithm
The previous commit would cut down *every* column proportionally until
we fit within the terminal width. However, this had the meaningful
downside of cutting 1-2 chars off small columns to make room for large
columns, which was really bad value in terms of almost pointlessly
hurting the readability of the small columns.

This commit changes the algo slightly, 'entitling' every column to at
least their fair share of all the termWidth, so no column below their
entitlement will get cut down, and instead the columns over their
entitlement will get cut down *extra*.

I dunno if this is like the 'optimal' algo for this sort of thing (I
should read how others solve this problem...) but it works well in my
head and in practice, from a bit of testing.

One small note: we simply add elipses to truncated columns. For columns
that are multiline with words, we should probably just re-wrap them and
make them longer. But it's a bit of an edge case.