Add debug print test + fixes
Working out the testing framework a bit more, found some things I needed
to do to make the testing work correctly (especially around resetting
rootModified).

Also simplify the creation of the rootCmd a bit by not relying on an
init function.

Kinda a happy accident with writing this as my second test -- I already
found what I think is overlooked behavior on ScriptDebug and RadDebug,
in terms of them *always* writing to Stderr, which we probably don't
want -- only if SHELL mode is active.