Don't fail if unable to get terminal width
I'm not 100% confident in when it happens, but I've observed it in the
a few of the ./run_tests.sh tests. But the embedded RSL example I'm
adding in this commit, it seems to work fine 🤷

Anyway, either way, it's not a massive deal. Let's just say 9999 width,
which will in practice uncap the width of the table we're willing to
print, just printing it 'raw' according to however much width the data
needs.