Implement compound assignments
i.e.

a = 1
a += 2

a will be 3.

Included compound assignment operators are

+= ints, floats, strings, arrays
-= ints, floats
*= ints, floats
/= ints, floats

We evolve our testing a bit more in this commit. Slowly figuring it out
:) We split the test files a bit, and implement a 'paradigm' for
embedding the RSL code as a block string in the test itself and passing
it in via the test Stdin buffer, rather than making a .rad file and
pointing the test at it. This is nice because it keeps the test input
local to the test, and lessens the number of files we need. There might
still be a place for standalone files for long rad scripts, though
arguably that could totally still be embedded. We'll see.

While here, we also actually fix some bugs for our math. Not sure how I
missed it, but adding e.g. a = 1 + 1.1 would not work before, but now it
will. The issue was trying to convert 1 into a float by doing
var.(float64), rather than float64(var.(int)).