Implement --MOCK-RESPONSE <urlRegex:jsonFilePath> flag
This will server two use cases:

1. It allows us to unit test more request-related and table-printing
functionality without actually hitting an external API, which is great
to not get throttled.

2. It allows RSL devs to test and write tests for their own scripts,
changing the responses they might get from the endpoint they're hitting.

Doing this 'MockResponseSlice' approach allows users to configure the
flag multiple times to set different mock responses for different URL
regex matches.

Doesn't serve the case where you want to *change* the mocked response
over time. But we can cross that bridge if we get to it.

Other changes:

To aid debugging, we print all args to the program if --RAD-DEBUG is
set.

Fix a small edge case in the table shrinking logic. Noticed that we
could negatively index one of the arrays when truncating.