Do not use exponential notation for long numbers in table
Long numbers would get printed as e.g. 1.1234e+16 rather than the more
'normal' form. This is a Go implementation detail, not intended behavior
for rad.

We add a test to check this, which does reveal that we actually hit Go's
float64 precision issues with large numbers in JSON. Maybe there's a way
to tell Go to treat all numbers as strings, even if they're not
specified that way, just to maintain precision? Is that good though?

Idk, something for another day.

Also discovered the wildcard extraction is nondeterministic when it
comes to ordering of the table, giving flaky tests. It should be
deterministic, will need to take a look at that. Probably related to
iterating through maps?