Replace olekukonko/tablewriter with amterp/go-tbl
I was working on colored headers and noticed a bug which made them not
line up properly with the rows below. This seems like a tablewriter bug
but this library is not actively maintained.

I foresee a few of these things will pop up. rad will depend quite
heavily on its table writing library so I think we should just take the
step to fork the original now so I can make the modifications I want.