Colorize and uncapitalize table headers
Prior to this commit, tables are printed as

NAME   AGE
Alice  30
Bob    25

regardless of the variable name for the json fields name and age.

This was a compromise for making headers visible before we implement
coloring, but with this commit, we implement coloring the header. This
is currently hardcoded to yellow, but we will probably add some RSL
syntax down the road (or rad config) to change it.

The result is that if your json fields are named Name and Age, you get
this:

Name   Age           << imagine this is yellow :)
Alice  30
Bob    25

This depends on some changes made in go-tbl which leverage fatih/color,
which has the benefit of handling coloring for cases where the rad
output is piped to a file for example, which we don't want coloring for
(this will just show up as ANSI escape codes).

We add a global rad flag --NO-COLOR to disable all coloring.