Use int64 instead of int to represent integers
Shoulda done this from the start, it just lowers the risk of overflow so
we can be more likely to maintain correctness of our scripts, and
performance (both in mem and speed) is not really a concern for rad's
usecase.

We also move one of the 'bash' date function tests to proper Go testing.
This requires being able to mock out the clock / source of time for rad,
hence the Clock being added in this commit.