Bug fix: --SHELL with -h would segfault on RP being nil
Previously, RP would only get defined in 'PersistentPreRun' of the root
cmd. If -h was passed, this would not run before the help function, and
so the 'exit 0' print would fail as RP is undefined.

Let's fix that by creating RP inside the help function as well,
optionally. We need to ResetGlobals (it's more correct) for resetting
between tests, but it's particularly required becuase of the RP == nil
check we add.