Sort JSON map keys prior to iterating for wildcard traversal
We currently have some flaky tests caused by non deterministic behavior.
It steps from iterating through the map[string]interface{} when
traversing for a wildcard.

We *want* to maintain consistent behavior with the rest of the json algo
guarantees, in that, unless the rad block has a 'sort', it will print
extractions in the order that they appear in the JSON. This includes
maps. However, this is a bigger lift that I would rather revisit later.

I'll let the todo I've added in the code speak a bit more:

 at this point this is a concession -- we should be traversing in the
 original order of the json but the json.Unmarshalling we're doing loses
 us the original ordering for maps. will need to change how we get the
 json if we want to change that (we almost certainly do, this is
 idiosyncratic behavior).

We opt to just sort the keys before traversing. At least it gives us
deterministic behavior. We fix the associated test to reflect this.

To try and avoid more flaky tests in the future, I add a push.sh script
which should be used before pushing. This will rebuild the project and
run all tests 50 times, and only then allow pushing.