Implement 'pick' function
This function allows choosing a single element as output given an array
of strings.

 options = ["Hamburger", "Chicken Burger", "Sandwich"]
 out = pick(options)
 print(out) // prints e.g. 'Hamburger', if that's what got picked

This will launch rad into a interactive CLI which lets the user pick an
element (and filter down further if they want).

A second argument can be given to do an initial programmatic filtering
of the options. If there's just one remaining option, then that gets
picked. If there are still multiple, we again do the 'interactive'
select.

 args:
     filter string = ""

 options = ["Hamburger", "Chicken Burger", "Sandwich"]
 out = pick(options, filter)
 print(out)

The above example allows rad users to provide their own prefilter, if
they want.

The interactive picking is executed by charmbracelet/huh. It looks neat,
though on the surface it also looks quite heavy as a dependency. Just
adding it to rad has almost tripled the number of lines in go.sum. We
might revisit if that's a problem.

We try to write some unit tests but I omit ones that launch the
interactive CLI, since I'm not 100% sure at this point how we'd do that
from a unit test. Would in theory need to be able to pass some second
stdin or request additional stdin that the unit test provides (this is
complicated by us already using stdin for passing rsl to the cmd).

I want to allow RSL devs to write a prompt that gets used by huh, but
that will require the implementation of named args in RSL, which will
come later.

I add a small markdown on my thoughts for resources atm. Thinking simple
json files might not be a bad bet, rather than another custom syntax.