Support multi-assignment for functions
This is required to allow pick_from_resource in a future commit, where
we want functions to be able to return multiple values and assign each
value to a different variable.

We don't allow other forms of multi-assignment at this point, there's no
syntax for e.g. having the comma between the two numbers here:

 a, b = 1, 2

Maybe one day, perhaps it's nice for consistency with pick_from_resource
or switch assignments, but not for now.