Allow string[] of filters in pick, pick_kv
To allow fuzzy matching against several tokens. e.g. [Ham, ger] will
match "Hamburger".

We do this to both pick and pick_kv, but leave pick_with_resource as-is,
since it's not a 'fuzzy' match in the same way as the others right now,
but rather a lookup, effectively. Once we make PWR fuzzy as well though,
this change would also make sense here.