pick_from_resource: change schema to 'key/values', not 'match/return'
The latter is problematic in part because 'return' is a keyword in a lot
of languages, so if you want to use e.g. Python to generate rad resource
files, this naming can be an obstacle depending on how to you do it.

Let's just name it 'keys' and 'values'. It's more aligned with other
methods e.g. pick_kv and is also perhaps a bit more intuitive.