Rework arrays: allow 'mixed' arrays, including arrays within arrays
The typed arrays approach is good for *arguments*, and it allows for
some nice API controls at a script-to-user level.

However, when it comes to implementing the RSL script itself, it can be
a bit limiting. We really want to be able to support arrays within
arrays. One reason is that it might pop up in JSON that an RSL dev wants
to handle in a request. By greatly improving the usability of arrays and
allowing mixed types, we support a much broader schema of JSONs and rad
as a whole becomes more flexible.

It also helps to open us up to some neat features down the road,
particularly Python-style list comprehensions. For example, being able
to print the lengths of captured arrays in a table could be done like
so:

 arrays = json[].ids  // captures as an array of arrays
 request url:
    fields arrays

 lens = [len(arr) for arr in arrays]  // list comprehension for lengths

 display:
    fields lens

I've hopefully remembered to update all the relevant locations to add
mixed array support, particularly for functions. Might've missed some
spots, will fix as they pop up, and add the appropriate test coverage
that'd imply we're missing.

===

Since we're touching this whole area of the code, we also take this
opportunity to be a bit more consistent in our bool array support. I've
only been half-implementing it as I've not been terribly convinced of
its usefulness, but this half-baked state is probably worse. There
*might* be some use for it, and if nothing else, it'd be inconsistent
to allow strings, ints, floats, bools, and arrays of all those *except*
bools.