Rework array access/indexing
Prior, our array access grammar was very limited. It explicitly required
an identifier and then would allow a single index lookup. This was
mostly okay as we didn't allow arrays within arrays, though wouldn't
have allowed e.g. myFunction()[0] where myFunction returned an array,
which is not ideal.

Now that arrays within arrays *are* implemented however, this issue
becomes much more pertinent. Let's rework it a little so that instead
of our grammar being

 arrayAccess  ->  IDENTIFIER [ EXPR ]

it is now

 arrayAccess  ->  PRIMARY [ EXPR ]

plus, in the parser, we won't stop at reading one index access, we'll
keep index accesses and wrap each in a visitable ArrayAccess AST node so
that our interpreter can correctly visit the array accesses and we get
the returned values that we expect.

On the ArrayAccess AST type, we include a 'OpenBracketToken' just so
that we have something to point to in our parser errors, if needed.