Implement list (array) comprehensions
Technically we've been referring to arrays as arrays, not lists, but
'list comprehension' is probably more familiar to devs so opting for
that terminology here.

The syntax is lifted straight from Python, which has two benefits:

1. It's familiar for devs so makes RSL easier to pick up
2. It fits well with existing RSL syntax, e.g. for .. in is also how
   we do for each loops.

We include the 'filtering' version as well which allows tacking on an
if statement at the end to filter if the expression gets evaluated or
not.

One thing lacking, that Python allows and I certainly want to as well in
RSL, is allowing e.g. 'print' statements in the list comprehension. It's
tricky, and we don't do it here, because I've previously made the
decision to not have nulls or a null-like type in RSL, and so void
functions like 'print' truly don't return anything, and cannot be used
in places where a return value is expected (i.e. we only allow them as
statements). However, list comprehensions that do prints are pretty
useful; they're a concise way to print something in a certain way, for
example:

 names = ["Alice", "Bob", "Charlie"]
 for name in names:
     print("Hello ", name)

can be replaced with

 names = ["Alice", "Bob", "Charlie"]
 [print("Hello ", name) for name in names]

Not a big loss, as you can see though, if we don't support it. I might
revisit this if I think of a nice way to handle it.

Misc bug fix: I added a test 'TestCanUseVarsInArrays' which demonstrates
the issue, but I was inadvertently forcing all array definitions to be
array literals, meaning 'expressions' were not allowed. Something like
this was not allowed:

 "a"
 b = 1
 c = true
 print([a, b, c])

it would fail on the [a, b, c]. Anyway, that's fixed.