Add 'request' and 'display' blocks
This is a very useful change -- currently, the only block for extracting
fields and printing tables is the 'rad' block. This is somewhat
appropriately named because it does both 'request and display'. However,
there are many use cases where you might want to do only one or the
other. For example, you might need to first do a request, extracting
some data, without printing out the raw values as captured, so that you
can then do some more processing with the captured data, before either
doing another request and then display, or going straight to a display
action.

For example, we have the below unit test to demo. The data it operates
on is this:

  [
    {
      "name": "Alice",
      "ids": [1, 2, 3]
    },
    {
      "name": "Bob",
      "ids": [4, 5, 6, 7, 8]
    },
    {
      "name": "Charlie",
      "ids": [9, 10]
    }
  ]

and here is the RSL:

  url = "some-url.com"

  Name = json[].name
  ids = json[].ids

  request url:
      fields Name, ids

  NumIds = [len(x) for x in ids]

  display:
      fields Name, NumIds

Here you can see that we first do a 'request', which gives us the names
and the 'ids' arrays. However, we're not interested in displaying the
ids themselves, we just want to display the *number* of IDs that each
person has. Previously, we couldn't do this, because the 'rad' block
would immediately print the ids. Now, however, we can use this 'request'
block which does not print any captured data or table, and we can then
perform a list comprehension to get the lengths of the ids for each
person. Then, we can use the resulting array together with the original
'Names' array, to print the final output:

  Name     NumIds
  Alice    3
  Bob      5
  Charlie  2

Implementation-wise, request/display/rad blocks are all really a 'rad
block' underneath the hood, with different fields set and validation
applied. This is probably the simplest and least duplication-heavy
approach to this, given the venn diagram of overlapping stmts that would
apply to each type of block.

One thing we needed to change in the json algo was not converting every
captured piece of data into a string with ToPrintable. Now we instead
leave it as-is (interface{}) which maintains their underlying datatype.
This is required so that e.g. the 'ids' array in the above example
doesn't simply become a string, but remains an array, which is important
for further data manipulation. Thankfully a painless change to make, no
tests broke.